some logical corrections
